+++ config/aggregate_test.go
@@ -33,8 +33,8 @@ func TestAggregate(t *testing.T) {
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "19cb98d5-dd17-4daf-ae00-1b413b7b899a")
 						c.Routes(
-							dogma.HandlesCommand[CommandStub[TypeA]](),
-							dogma.RecordsEvent[EventStub[TypeA]](),
+							dogma.HandlesCommand[*CommandStub[TypeA]](),
+							dogma.RecordsEvent[*EventStub[TypeA]](),
 						)
 					},
 				}),
@@ -133,8 +133,8 @@ func TestAggregate(t *testing.T) {
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "non-uuid")
 						c.Routes(
-							dogma.HandlesCommand[CommandStub[TypeA]](),
-							dogma.RecordsEvent[EventStub[TypeA]](),
+							dogma.HandlesCommand[*CommandStub[TypeA]](),
+							dogma.RecordsEvent[*EventStub[TypeA]](),
 						)
 					},
 				}),
@@ -148,8 +148,8 @@ func TestAggregate(t *testing.T) {
 						c.Identity("foo", "63bd2756-2397-4cae-b33b-96e809b384d8")
 						c.Identity("bar", "ee316cdb-894c-454e-91dd-ec0cc4531c42")
 						c.Routes(
-							dogma.HandlesCommand[CommandStub[TypeA]](),
-							dogma.RecordsEvent[EventStub[TypeA]](),
+							dogma.HandlesCommand[*CommandStub[TypeA]](),
+							dogma.RecordsEvent[*EventStub[TypeA]](),
 						)
 					},
 				}),
@@ -162,7 +162,7 @@ func TestAggregate(t *testing.T) {
 						c.Identity("handler", "d1e04684-ec56-44a7-8c7d-f111b2d6b2d2")
 						c.Routes(
 							// <-- MISSING HandlesCommand ROUTE
-							dogma.RecordsEvent[EventStub[TypeA]](),
+							dogma.RecordsEvent[*EventStub[TypeA]](),
 						)
 					},
 				}),
@@ -174,7 +174,7 @@ func TestAggregate(t *testing.T) {
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("handler", "d1e04684-ec56-44a7-8c7d-f111b2d6b2d2")
 						c.Routes(
-							dogma.HandlesCommand[CommandStub[TypeA]](),
+							dogma.HandlesCommand[*CommandStub[TypeA]](),
 							// <-- MISSING "RecordEvent" ROUTE
 						)
 					},
@@ -182,28 +182,28 @@ func TestAggregate(t *testing.T) {
 			},
 			{
 				Name:  "aggregate must not have multiple routes for the same command type",
-				Error: `aggregate:AggregateMessageHandlerStub is invalid: multiple handles-command routes for github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+				Error: `aggregate:AggregateMessageHandlerStub is invalid: multiple handles-command routes for *github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("handler", "d1e04684-ec56-44a7-8c7d-f111b2d6b2d2")
 						c.Routes(
-							dogma.HandlesCommand[CommandStub[TypeA]](), // <-- SAME MESSAGE TYPE
-							dogma.HandlesCommand[CommandStub[TypeA]](), // <-- SAME MESSAGE TYPE
-							dogma.RecordsEvent[EventStub[TypeA]](),
+							dogma.HandlesCommand[*CommandStub[TypeA]](), // <-- SAME MESSAGE TYPE
+							dogma.HandlesCommand[*CommandStub[TypeA]](), // <-- SAME MESSAGE TYPE
+							dogma.RecordsEvent[*EventStub[TypeA]](),
 						)
 					},
 				}),
 			},
 			{
 				Name:  "aggregate must not have multiple routes for the same event type",
-				Error: `aggregate:AggregateMessageHandlerStub is invalid: multiple records-event routes for github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+				Error: `aggregate:AggregateMessageHandlerStub is invalid: multiple records-event routes for *github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("handler", "d1e04684-ec56-44a7-8c7d-f111b2d6b2d2")
 						c.Routes(
-							dogma.HandlesCommand[CommandStub[TypeA]](),
-							dogma.RecordsEvent[EventStub[TypeA]](), // <-- SAME MESSAGE TYPE
-							dogma.RecordsEvent[EventStub[TypeA]](), // <-- SAME MESSAGE TYPE
+							dogma.HandlesCommand[*CommandStub[TypeA]](),
+							dogma.RecordsEvent[*EventStub[TypeA]](), // <-- SAME MESSAGE TYPE
+							dogma.RecordsEvent[*EventStub[TypeA]](), // <-- SAME MESSAGE TYPE
 						)
 					},
 				}),
@@ -253,15 +253,15 @@ func TestAggregate(t *testing.T) {
 				Description: multiline(
 					`valid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub`,
 					`  - valid identity name/19cb98d5-dd17-4daf-ae00-1b413b7b899a`,
-					`  - valid handles-command route for github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
-					`  - valid records-event route for github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+					`  - valid handles-command route for *github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+					`  - valid records-event route for *github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 				),
 				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "19cb98d5-dd17-4daf-ae00-1b413b7b899a")
 						c.Routes(
-							dogma.HandlesCommand[CommandStub[TypeA]](),
-							dogma.RecordsEvent[EventStub[TypeA]](),
+							dogma.HandlesCommand[*CommandStub[TypeA]](),
+							dogma.RecordsEvent[*EventStub[TypeA]](),
 						)
 					},
 				}),
@@ -272,16 +272,16 @@ func TestAggregate(t *testing.T) {
 				Description: multiline(
 					`valid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub`,
 					`  - valid identity name/19cb98d5-dd17-4daf-ae00-1b413b7b899a`,
-					`  - valid handles-command route for github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
-					`  - valid records-event route for github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+					`  - valid handles-command route for *github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+					`  - valid records-event route for *github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 					`  - valid disabled flag, set to true`,
 				),
 				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "19cb98d5-dd17-4daf-ae00-1b413b7b899a")
 						c.Routes(
-							dogma.HandlesCommand[CommandStub[TypeA]](),
-							dogma.RecordsEvent[EventStub[TypeA]](),
+							dogma.HandlesCommand[*CommandStub[TypeA]](),
+							dogma.RecordsEvent[*EventStub[TypeA]](),
 						)
 						c.Disable()
 					},
@@ -345,15 +345,15 @@ func TestAggregate(t *testing.T) {
 					`valid aggregate *github.com/dogmatiq/enginekit/enginetest/stubs.AggregateMessageHandlerStub`,
 					`  - invalid identity name/non-uuid`,
 					`      - invalid key ("non-uuid"), expected an RFC 4122/9562 UUID`,
-					`  - valid handles-command route for github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
-					`  - valid records-event route for github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+					`  - valid handles-command route for *github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+					`  - valid records-event route for *github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 				),
 				Component: runtimeconfig.FromAggregate(&AggregateMessageHandlerStub{
 					ConfigureFunc: func(c dogma.AggregateConfigurer) {
 						c.Identity("name", "non-uuid")
 						c.Routes(
-							dogma.HandlesCommand[CommandStub[TypeA]](),
-							dogma.RecordsEvent[EventStub[TypeA]](),
+							dogma.HandlesCommand[*CommandStub[TypeA]](),
+							dogma.RecordsEvent[*EventStub[TypeA]](),
 						)
 					},
 				}),
@@ -367,8 +367,8 @@ func TestAggregate(t *testing.T) {
 				ConfigureFunc: func(c dogma.AggregateConfigurer) {
 					c.Identity("name", "19cb98d5-dd17-4daf-ae00-1b413b7b899a")
 					c.Routes(
-						dogma.HandlesCommand[CommandStub[TypeA]](),
-						dogma.RecordsEvent[EventStub[TypeA]](),
+						dogma.HandlesCommand[*CommandStub[TypeA]](),
+						dogma.RecordsEvent[*EventStub[TypeA]](),
 					)
 				},
 			}
@@ -380,8 +380,8 @@ func TestAggregate(t *testing.T) {
 				"unexpected routes",
 				handler.RouteSet().MessageTypes(),
 				map[message.Type]RouteDirection{
-					message.TypeFor[CommandStub[TypeA]](): InboundDirection,
-					message.TypeFor[EventStub[TypeA]]():   OutboundDirection,
+					message.TypeFor[*CommandStub[TypeA]](): InboundDirection,
+					message.TypeFor[*EventStub[TypeA]]():   OutboundDirection,
 				},
 			)
 		})
@@ -399,32 +399,32 @@ func TestAggregate(t *testing.T) {
 				},
 				{
 					"unsupported ExecutesCommand route",
-					`unsupported executes-command route for github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+					`unsupported executes-command route for *github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 					&Route{
 						RouteType:       optional.Some(ExecutesCommandRouteType),
-						MessageTypeID:   optional.Some(MessageTypeID[CommandStub[TypeA]]()),
-						MessageTypeName: optional.Some("github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
-						MessageType:     optional.Some(message.TypeFor[CommandStub[TypeA]]()),
+						MessageTypeID:   optional.Some(MessageTypeID[*CommandStub[TypeA]]()),
+						MessageTypeName: optional.Some("*github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
+						MessageType:     optional.Some(message.TypeFor[*CommandStub[TypeA]]()),
 					},
 				},
 				{
 					"unsupported HandlesEvent route",
-					`unsupported handles-event route for github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+					`unsupported handles-event route for *github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 					&Route{
 						RouteType:       optional.Some(HandlesEventRouteType),
-						MessageTypeID:   optional.Some(MessageTypeID[EventStub[TypeA]]()),
-						MessageTypeName: optional.Some("github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
-						MessageType:     optional.Some(message.TypeFor[EventStub[TypeA]]()),
+						MessageTypeID:   optional.Some(MessageTypeID[*EventStub[TypeA]]()),
+						MessageTypeName: optional.Some("*github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
+						MessageType:     optional.Some(message.TypeFor[*EventStub[TypeA]]()),
 					},
 				},
 				{
 					"unsupported SchedulesTimeout route",
-					`unsupported schedules-timeout route for github.com/dogmatiq/enginekit/enginetest/stubs.TimeoutStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
+					`unsupported schedules-timeout route for *github.com/dogmatiq/enginekit/enginetest/stubs.TimeoutStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]`,
 					&Route{
 						RouteType:       optional.Some(SchedulesTimeoutRouteType),
-						MessageTypeID:   optional.Some(MessageTypeID[TimeoutStub[TypeA]]()),
-						MessageTypeName: optional.Some("github.com/dogmatiq/enginekit/enginetest/stubs.TimeoutStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
-						MessageType:     optional.Some(message.TypeFor[TimeoutStub[TypeA]]()),
+						MessageTypeID:   optional.Some(MessageTypeID[*TimeoutStub[TypeA]]()),
+						MessageTypeName: optional.Some("*github.com/dogmatiq/enginekit/enginetest/stubs.TimeoutStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
+						MessageType:     optional.Some(message.TypeFor[*TimeoutStub[TypeA]]()),
 					},
 				},
 			}