+++ config/runtimeconfig/aggregate_test.go
@@ -55,8 +55,8 @@ func TestFromAggregate(t *testing.T) {
 				ConfigureFunc: func(c dogma.AggregateConfigurer) {
 					c.Identity("aggregate", "d9d75a75-7839-4b3e-a7e5-c8884b88ea57")
 					c.Routes(
-						dogma.HandlesCommand[CommandStub[TypeA]](),
-						dogma.RecordsEvent[EventStub[TypeA]](),
+						dogma.HandlesCommand[*CommandStub[TypeA]](),
+						dogma.RecordsEvent[*EventStub[TypeA]](),
 					)
 					c.Disable()
 				},
@@ -76,15 +76,15 @@ func TestFromAggregate(t *testing.T) {
 						RouteComponents: []*config.Route{
 							{
 								RouteType:       optional.Some(config.HandlesCommandRouteType),
-								MessageTypeID:   optional.Some(MessageTypeID[CommandStub[TypeA]]()),
-								MessageTypeName: optional.Some("github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
-								MessageType:     optional.Some(message.TypeFor[CommandStub[TypeA]]()),
+								MessageTypeID:   optional.Some(MessageTypeID[*CommandStub[TypeA]]()),
+								MessageTypeName: optional.Some("*github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
+								MessageType:     optional.Some(message.TypeFor[*CommandStub[TypeA]]()),
 							},
 							{
 								RouteType:       optional.Some(config.RecordsEventRouteType),
-								MessageTypeID:   optional.Some(MessageTypeID[EventStub[TypeA]]()),
-								MessageTypeName: optional.Some("github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
-								MessageType:     optional.Some(message.TypeFor[EventStub[TypeA]]()),
+								MessageTypeID:   optional.Some(MessageTypeID[*EventStub[TypeA]]()),
+								MessageTypeName: optional.Some("*github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
+								MessageType:     optional.Some(message.TypeFor[*EventStub[TypeA]]()),
 							},
 						},
 						DisabledFlags: []*config.Flag[config.Disabled]{