+++ config/runtimeconfig/process_test.go
@@ -53,9 +53,9 @@ func TestFromProcess(t *testing.T) {
 				ConfigureFunc: func(c dogma.ProcessConfigurer) {
 					c.Identity("projection", "050415ad-ce90-496f-8987-40467e5415e0")
 					c.Routes(
-						dogma.HandlesEvent[EventStub[TypeA]](),
-						dogma.ExecutesCommand[CommandStub[TypeA]](),
-						dogma.SchedulesTimeout[TimeoutStub[TypeA]](),
+						dogma.HandlesEvent[*EventStub[TypeA]](),
+						dogma.ExecutesCommand[*CommandStub[TypeA]](),
+						dogma.SchedulesTimeout[*TimeoutStub[TypeA]](),
 					)
 					c.Disable()
 				},
@@ -75,21 +75,21 @@ func TestFromProcess(t *testing.T) {
 						RouteComponents: []*config.Route{
 							{
 								RouteType:       optional.Some(config.HandlesEventRouteType),
-								MessageTypeID:   optional.Some(MessageTypeID[EventStub[TypeA]]()),
-								MessageTypeName: optional.Some("github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
-								MessageType:     optional.Some(message.TypeFor[EventStub[TypeA]]()),
+								MessageTypeID:   optional.Some(MessageTypeID[*EventStub[TypeA]]()),
+								MessageTypeName: optional.Some("*github.com/dogmatiq/enginekit/enginetest/stubs.EventStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
+								MessageType:     optional.Some(message.TypeFor[*EventStub[TypeA]]()),
 							},
 							{
 								RouteType:       optional.Some(config.ExecutesCommandRouteType),
-								MessageTypeID:   optional.Some(MessageTypeID[CommandStub[TypeA]]()),
-								MessageTypeName: optional.Some("github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
-								MessageType:     optional.Some(message.TypeFor[CommandStub[TypeA]]()),
+								MessageTypeID:   optional.Some(MessageTypeID[*CommandStub[TypeA]]()),
+								MessageTypeName: optional.Some("*github.com/dogmatiq/enginekit/enginetest/stubs.CommandStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
+								MessageType:     optional.Some(message.TypeFor[*CommandStub[TypeA]]()),
 							},
 							{
 								RouteType:       optional.Some(config.SchedulesTimeoutRouteType),
-								MessageTypeID:   optional.Some(MessageTypeID[TimeoutStub[TypeA]]()),
-								MessageTypeName: optional.Some("github.com/dogmatiq/enginekit/enginetest/stubs.TimeoutStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
-								MessageType:     optional.Some(message.TypeFor[TimeoutStub[TypeA]]()),
+								MessageTypeID:   optional.Some(MessageTypeID[*TimeoutStub[TypeA]]()),
+								MessageTypeName: optional.Some("*github.com/dogmatiq/enginekit/enginetest/stubs.TimeoutStub[github.com/dogmatiq/enginekit/enginetest/stubs.TypeA]"),
+								MessageType:     optional.Some(message.TypeFor[*TimeoutStub[TypeA]]()),
 							},
 						},
 						DisabledFlags: []*config.Flag[config.Disabled]{