+++ marshaler/codecs/protobuf/doc.go
@@ -1,3 +1,0 @@
-// Package protobuf contains codec implementations for protocol buffers
-// types.
-package protobuf