+++ marshaler/codecs/stateless/doc.go
@@ -1,3 +1,0 @@
-// Package stateless is a codec that provides efficient encoding of
-// [dogma.StatelessProcessRoot] values.
-package stateless