+++ marshaler/doc.go
@@ -1,4 +1,0 @@
-// Package marshaler provides abstractions for marshaling and unmarshaling
-// values to and from a "portable" representation for persistence and
-// transmission over networks.
-package marshaler