+++ message/interface.go
@@ -38,23 +38,17 @@ func interfaceOf(r reflect.Type) reflect.Type {
 		panic("message type must not be nil")
 	}
 
-	if i := tryInterfaceOf(r); i != nil {
-		if r.Kind() == reflect.Pointer {
-			e := r.Elem()
-			if i := tryInterfaceOf(e); i != nil {
-				panic(fmt.Sprintf("%s does not use a pointer receiver to implement %s, use %s instead", r, i, e))
-			}
-		}
-
-		return i
+	i := tryInterfaceOf(r)
+	if i == nil {
+		panic(fmt.Sprintf("%s does not implement dogma.Command, dogma.Event or dogma.Timeout", r))
 	}
 
-	p := reflect.PointerTo(r)
-	if i := tryInterfaceOf(p); i != nil {
-		panic(fmt.Sprintf("%s uses a pointer receiver to implement %s, use %s instead", r, i, p))
+	if r.Kind() != reflect.Pointer {
+		p := reflect.PointerTo(r)
+		panic(fmt.Sprintf("%s implements %s, but message implementations must use pointer receivers, use %s instead", r, i, p))
 	}
 
-	panic(fmt.Sprintf("%s does not implement dogma.Command, dogma.Event or dogma.Timeout", r))
+	return i
 }
 
 func guardAgainstNonMessage(r reflect.Type) {